Spring data

+++ src/main/java/ru/clevertec/data/repository/CrudRepository.java
@@ -5,11 +5,11 @@ import ru.clevertec.util.exceptions.RepositoryException;
 import java.util.List;
 
 public interface CrudRepository <T>{
-    T getById(int id) throws RepositoryException;
+    T getById(Long id) throws RepositoryException;
 
     T add(T item) throws RepositoryException;
 
-    boolean remove(Integer id) throws RepositoryException;
+    boolean remove(Long id) throws RepositoryException;
 
     boolean update(T product) throws RepositoryException;
